refactor(standby_monitor): split HA heartbeat + promotion into submodules
standby_monitor.rs was 2289 lines (~1384 production): the two supervised HA
loops — the primary heartbeat writer and the standby promotion monitor — plus
their shared config/identity vocabulary and three test suites, all in one file.
Convert to a standby_monitor/ module dir and move each loop to its concern
submodule:

- standby_monitor/heartbeat.rs (352) — spawn_heartbeat_writer + heartbeat_loop
  + HeartbeatFreshness + the promotion_decision freshness gate.
- standby_monitor/promotion.rs (704) — OnPromote + spawn_promotion_monitor +
  apply_post_promotion + promotion_loop + perform_promotion.
- standby_monitor/mod.rs — the shared vocabulary (timing constants, instance
  identity: init/instance/resolve_instance_id, HaTimings, the state keys, the
  pure decision helpers) + the three test modules.

mod.rs production drops ~1384 -> ~349 lines. Behaviour is unchanged — a pure
relocation. The public entry points (spawn_heartbeat_writer,
spawn_promotion_monitor, OnPromote) are re-exported from mod.rs, so external
paths (env_config, lease, the service bin) are stable. The two shared state
keys became pub(crate); HeartbeatFreshness/promotion_decision (used across the
split and by the tests) are re-exported pub(crate); the test-only promotion
internals via a #[cfg(test)] re-export. forward-internal fns stay private to
their concern module.

kirra-verifier: lib + bins build; the 3 moved test modules pass (10 + 5 + 16);
clippy clean; fmt clean. De-monolith #4 (medium tier), after #978/#979/#980.
